+++ _define.php
@@ -17,27 +17,14 @@ $this->registerModule(
     'Mail to log',
     'Do not send mails but log them',
     'Jean-Christian Denis and contributors',
-    '0.7.1',
+    '0.8',
     [
         'requires'    => [['core', '2.28']],
         'permissions' => 'My',
+        'priority'    => 10,
         'type'        => 'plugin',
         'support'     => 'https://git.dotclear.watch/JcDenis/' . basename(__DIR__) . '/issues',
         'details'     => 'https://git.dotclear.watch/JcDenis/' . basename(__DIR__) . '/src/branch/master/README.md',
         'repository'  => 'https://git.dotclear.watch/JcDenis/' . basename(__DIR__) . '/raw/branch/master/dcstore.xml',
     ]
 );
-
-if (!function_exists('\_mail')) {
-    function _mail(string $to, string $subject, string $message, $headers = '', ?string $params = null): bool
-    {
-        $headers = is_array($headers) ? implode("\n\t", $headers) : (is_string($headers) ? $headers : '');
-
-        $cur = Dotclear\App::log()->openLogCursor();
-        $cur->setField('log_table', basename(__DIR__));
-        $cur->setField('log_msg', sprintf("%s\n-----\n To: %s\n Subject: %s\n-----\n Message:\n%s\n", $headers, $to, $subject, $message));
-        Dotclear\App::log()->addLog($cur);
-
-        return true;
-    }
-}