+++ src/My.php
@@ -7,8 +7,8 @@ namespace Dotclear\Plugin\mail2log;
 use Dotclear\Module\MyPlugin;
 
 /**
- * @brief   mail2log My helper.
- * @ingroup mail2log
+ * @brief       mail2log My helper.
+ * @ingroup     mail2log
  *
  * @author      Jean-Christian Denis (author)
  * @copyright   GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html