+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+mail2log 0.8 - 2023.10.23
+===========================================================
+* Require Dotclear 2.28
+* Require PHP 8.1
+* Move mail function to keep clean the define
+* Add plugin Uninstaller features
+
 mail2log 0.7.1 - 2023.10.19
 ===========================================================
 * Require Dotclear 2.28